refactor: reorganize package structure to follow naming conventions
