+++ convention-plugin/src/integrationTest/kotlin/io/github/aaravmahajanofficial/utils/BasicPluginConfiguration.kt
@@ -13,7 +13,7 @@
  * or implied. See the License for the specific language governing
  * permissions and limitations under the License.
  */
-package utils
+package io.github.aaravmahajanofficial.utils
 
 fun basicBuildScript(): String =
     """
