+++ convention-plugin/src/main/kotlin/io/github/aaravmahajanofficial/constants/PluginMetadata.kt
@@ -13,7 +13,7 @@
  * or implied. See the License for the specific language governing
  * permissions and limitations under the License.
  */
-package constants
+package io.github.aaravmahajanofficial.constants
 
 public object PluginMetadata {
     public const val EXTENSION_NAME: String = "jenkinsConvention"
