+++ convention-plugin/src/main/kotlin/io/github/aaravmahajanofficial/extensions/LicenseExtension.kt
@@ -13,7 +13,7 @@
  * or implied. See the License for the specific language governing
  * permissions and limitations under the License.
  */
-package extensions
+package io.github.aaravmahajanofficial.extensions
 
 import org.gradle.api.model.ObjectFactory
 import org.gradle.api.provider.ListProperty
