+++ convention-plugin/src/main/kotlin/io/github/aaravmahajanofficial/internal/GroovyConventionManager.kt
@@ -13,7 +13,7 @@
  * or implied. See the License for the specific language governing
  * permissions and limitations under the License.
  */
-package internal
+package io.github.aaravmahajanofficial.internal
 
 import org.gradle.api.Project
 import org.gradle.api.tasks.compile.GroovyCompile
