+++ convention-plugin/src/main/kotlin/io/github/aaravmahajanofficial/internal/JpiPluginAdapter.kt
@@ -13,9 +13,9 @@
  * or implied. See the License for the specific language governing
  * permissions and limitations under the License.
  */
-package internal
+package io.github.aaravmahajanofficial.internal
 
-import extensions.PluginExtension
+import io.github.aaravmahajanofficial.extensions.PluginExtension
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.getByType
 import org.gradle.kotlin.dsl.newInstance
