+++ convention-plugin/src/main/kotlin/io/github/aaravmahajanofficial/utils/GradleVersionUtils.kt
@@ -13,9 +13,9 @@
  * or implied. See the License for the specific language governing
  * permissions and limitations under the License.
  */
-package utils
+package io.github.aaravmahajanofficial.utils
 
-import constants.PluginMetadata
+import io.github.aaravmahajanofficial.constants.PluginMetadata
 import org.gradle.api.GradleException
 import org.gradle.util.GradleVersion
