+++ convention-plugin/src/main/kotlin/io/github/aaravmahajanofficial/utils/VersionCatalog.kt
@@ -13,7 +13,7 @@
  * or implied. See the License for the specific language governing
  * permissions and limitations under the License.
  */
-package utils
+package io.github.aaravmahajanofficial.utils
 
 import org.gradle.api.artifacts.MinimalExternalModuleDependency
 import org.gradle.api.artifacts.VersionCatalog
