+++ gradle.properties
@@ -1,7 +1,5 @@
 kotlin.code.style=official
 org.gradle.configuration-cache=true
-kotlin.daemon.jvmargs=-Xmx2048m
-org.gradle.jvmargs=-Xmx2048m
 org.jetbrains.dokka.experimental.gradle.pluginMode=V2Enabled
 org.jetbrains.dokka.experimental.gradle.pluginMode.noWarn=true
